// packages/ui/src/components/component.types.ts
export enum ColorTypes {
  DEFAULT = 'default',
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  TERTIARY = 'tertiary',
  ACTIVE = 'active',
  WARNING = 'warning',
  ALERT = 'alert',
  INFO = 'info',
}

export enum SizeTypes {
  SMALL = 'small',
  REGULAR = 'regular',
  LARGE = 'large',
}

export enum VariantTypes {
  SOLID = 'solid',
  FLAT = 'flat',
  OUTLINED = 'outlined',
  TEXT = 'text',
}

export enum RadiusTypes {
  NONE = 'none',
  SMALL = 'small',
  MEDIUM = 'medium',
  LARGE = 'large',
  FULL = 'full',
}